+++ tag-expressions/CHANGELOG.md
@@ -15,6 +15,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+* [Java] Updated `TagExpressionParser` to use a static method to parse a tag expression and return an `Expression` object to the user.
+* [Java] Reduced public API to the bare minimum required.
+* [Java] Added more informative error messages for `TagExpressionParser` through the `TagExpressionException`. 
+  ([#1005](https://github.com/cucumber/cucumber/pull/1005)
+  [cyocum])
+
 ### Deprecated
 
 ### Removed